irq.h: add __unused to the parameter of handle_nmi
In commit cca9e99a45fb ("reentrancy: update code and MDS according to
ACCB-194") the implementation of handle_nmi is modified in a way that the
given parameter is no longer necessary. The declaration of this function in
irq.h is, however, not updated accordingly.

This patch adds the __unused attribute to the only parameter of handle_nmi.

Tracked-On: ACCB-194
Signed-off-by: Junjie Mao <junjie.mao@intel.com>